backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-multipart
sentence-transformers
spacy
transformers
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
cors
python-docx
PyPDF2
scipy
requests
